Add files via upload
This is an example of an operating systems process management. It has an adding to priority queue method, a method for blocking said processes and a method of executing those processes according to each processes priority. Once we reach the final level of priority in this case queue 8. It will execute using the round robin method. It is fully  functional model in Python and displays the execution, blocking and priority change of each process in real time.